+++ src/main/java/com/mycompany/proyecto/persistencia/UsuarioJpaController.java
@@ -1,3 +1,4 @@
+
 package com.mycompany.proyecto.persistencia;
 
 import com.mycompany.proyecto.logica.Usuario;
@@ -22,7 +23,6 @@ public class UsuarioJpaController implements Serializable {
      public UsuarioJpaController(){
      emf= Persistence.createEntityManagerFactory("proJPAPU"); ;
     }
-    
     private EntityManagerFactory emf = null;
 
     public EntityManager getEntityManager() {
@@ -53,7 +53,7 @@ public class UsuarioJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                int id = usuario.getId();
+                Long id = usuario.getId();
                 if (findUsuario(id) == null) {
                     throw new NonexistentEntityException("The usuario with id " + id + " no longer exists.");
                 }
@@ -66,7 +66,7 @@ public class UsuarioJpaController implements Serializable {
         }
     }
 
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(Long id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -111,7 +111,7 @@ public class UsuarioJpaController implements Serializable {
         }
     }
 
-    public Usuario findUsuario(int id) {
+    public Usuario findUsuario(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Usuario.class, id);
